+++ CHANGELOG.md
@@ -30,6 +30,8 @@ x.y.z Release notes (yyyy-MM-dd)
 * Add EmailPasswordAuth flow implementation with Combine's Future
 * Add MongoCollection methods implementation with Combine's Future
 * Add support for User dynamicMemberLookup with Future result
+* Add Combine support in APIKeyAuth
+* Add Combine support in PushClient
 * Add ability to open a synced Realm with a `nil` partition value.
 
 ### Fixed